fix warnings; remove unnecessary linking dependencies
Former-commit-id: 8a616646cf97fd1aba1243784a5f8d84f92c2320

+++ apps/clients/mriclient/HDF5ImageAttribWriter.h
@@ -125,7 +125,7 @@ namespace Gadgetron
                 size_t ii;
                 for ( ii=0; ii<dim.size(); ii++ )
                 {
-                    dim2[ii] = dim[ii];
+                    dim2[ii] = (unsigned int)dim[ii];
                 }
 
                 if (dataset_.appendArray(dim2, data->get_data_ptr(), img_varname.c_str())  < 0)
